README.md: Fix comment about Gentoo
Citation needed; until then, lack of interest for packaging in official
repositories is no evidence of 'resistance' to do it. Downstream developers
have no duty to do so, any downstream developer would be free to add TDE as
long as package quality met the standards but no one ever stepped up.

However, the overlay, as it is, would have to undergo major improvements
before that could happen.

Signed-off-by: Andreas Sturmlechner <asturm@gentoo.org>