trinity-base: 9999 metapackages: set SLOT="14", set LICENSE="metapackage"
This is in preparation of cleaning up some bits of trinity-functions-2.eclass.

trinity-functions-2.eclass and set-trinityver do not really serve a purpose
here. The eclass is not meant to be inherited from individual ebuilds, and
if there ever will be a new SLOT it is going to be trivial to bump it.
SLOT usage should be revisited however, and possibly reverted to "0".

Also sort dependencies.

Signed-off-by: Andreas Sturmlechner <asturm@gentoo.org>