trinity-meta-2.eclass: Move SLOT/TDEDIR/TRINITY_VER setup to trinity-base-2
By making it conditional on CATEGORY=trinity-base there is effectively no
change for now with existing trinity-meta-2.eclass consumers.
This is a better fit, since all the things set up here are actually being
used inside trinity-base-2.eclass.

This is important for the next step, when the functions used are being
phased out in favor of meta vars.

In trinity-base category, this affects the following trinity-base-2.eclass
consumers:
- trinity-base/arts
- trinity-base/tde-i18n
- trinity-base/tdelibs

Of those, only trinity-base/tdelibs needs handling to not create a circular
dependency on its own.

Signed-off-by: Andreas Sturmlechner <asturm@gentoo.org>